# backend/docs/swagger.yaml
components:
  schemas:
...
    characters.CreateCharacterRequest:
      properties:
        abilities:
          type: string
        age:
          type: integer
        alias:
          type: string
        appearance:
          type: string
        avatar_url:
          type: string
        background_story:
          type: string
        gender:
          type: string
        name:
          type: string
        notes:
          type: string
        occupation:
          type: string
        personality:
          type: string
      required:
      - name
      type: object
...
    characters.UpdateCharacterRequest:
      properties:
        abilities:
          type: string
        age:
          type: integer
        alias:
          type: string
        appearance:
          type: string
        avatar_url:
          type: string
        background_story:
          type: string
        gender:
          type: string
        name:
          type: string
        notes:
          type: string
        occupation:
          type: string
        personality:
          type: string
      type: object
    data:
      properties:
        data:
          properties:
            message:
              type: string
          type: object
      type: object
...
    response.StandardResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      properties:
        code:
          type: integer
        data: {}
        message:
          type: string
        sourceId:
          type: string
        traceId:
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for Test.
  title: API 1.0
  version: "1.0"
openapi: 3.1.0
paths:
...
  /characters:
    get:
      description: Get a list of all characters with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to retrieve characters
      security:
      - BearerAuth: []
      summary: List all characters
      tags:
      - characters
    post:
      description: Create a new character with the given details
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/characters.CreateCharacterRequest'
        description: Create Character Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid request body
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to create character
      security:
      - BearerAuth: []
      summary: Create a new character
      tags:
      - characters
  /characters/{id}:
    delete:
      description: Delete a character by its ID
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Character not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to delete character
      security:
      - BearerAuth: []
      summary: Delete a character
      tags:
      - characters
    get:
      description: Get a single character by its ID
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Character not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to get character
      security:
      - BearerAuth: []
      summary: Get a single character
      tags:
      - characters
    put:
      description: Update a character with the given details
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/characters.UpdateCharacterRequest'
        description: Update Character Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID or request body
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Character not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to update character
      security:
      - BearerAuth: []
      summary: Update a character
      tags:
      - characters
...